restructured the env creation and specialization methods
